Mejoramientos al escripto de instalación.

+++ setup.py
@@ -1,20 +1,25 @@
 import os
 import platform
-import sys
-from subprocess import run
-import urllib.request
 import shutil
+import sys
+import urllib.request
+from subprocess import run
+from warnings import warn as avisar
+
 from setuptools import setup, find_packages
-
-
-with open('tinamit/versión.txt') as archivo_versión:
-    versión = archivo_versión.read().strip()
-
 
 """
 Esta parte instala todas las dependencias de Tinamit de manera automática. Al momento, funciona para Windows con
 Python 3.6 de 32 bits. Necesita una buena conexión internet.
 """
+
+versión_python = str(sys.version_info.major) + str(sys.version_info.minor)
+so = platform.system()
+bits = platform.architecture()[0][:2]
+
+directorio = os.path.split(os.path.realpath(__file__))[0]
+directorio_móds = os.path.join(directorio, 'Módulos')
+directorio_python = os.path.split(sys.executable)[0]
 
 try:
     import numpy as np
@@ -30,25 +35,6 @@ try:
     import scipy.stats as estad
 except ImportError:
     estad = None
-
-directorio = os.path.split(os.path.realpath(__file__))[0]
-
-directorio_móds = os.path.join(directorio, 'Módulos')
-
-if not os.path.exists(directorio_móds):
-    os.makedirs(directorio_móds)
-
-directorio_python = os.path.split(sys.executable)[0]
-
-versión_python = str(sys.version_info.major) + str(sys.version_info.minor)
-
-so = platform.system()
-
-if so != 'Windows':
-    raise OSError('Este programa de instalación funciona únicamente para Windows.')
-
-bits = platform.architecture()[0][:2]
-sistema = 'win' + bits
 
 info_paquetes = {'numpy': {'versión': '1.11.3',
                            'formato_archivo': 'numpy-{versión}+mkl-cp{v_py}-cp{v_py}m-{sis}.whl',
@@ -68,11 +54,6 @@ info_paquetes = {'numpy': {'versión': '1.11.3',
                  }
 
 
-for paquete, dic_paq in info_paquetes.items():
-    v = dic_paq['versión']
-    dic_paq['formato_archivo'] = dic_paq['formato_archivo'].format(versión=v, v_py=versión_python, sis=sistema)
-
-
 def _actualizar_pip():
     print('Actualizando pip...')
     comanda_pip = '%s install --upgrade pip' % (os.path.join(directorio_python, 'Scripts', 'pip'))
@@ -80,7 +61,6 @@ def _actualizar_pip():
 
 
 def _descargar_whl(nombre):
-
     print('Descargando paquete "{}"...'.format(nombre))
     llave = info_paquetes[nombre]['id_google'][bits]  # type: str
     if llave is None:
@@ -94,7 +74,6 @@ def _descargar_whl(nombre):
 
 
 def _instalar_whl(nombre):
-
     nombre_archivo = info_paquetes[nombre]['formato_archivo']
 
     if not os.path.isfile(os.path.join(directorio_móds, nombre_archivo)):
@@ -120,6 +99,13 @@ def instalar_requísitos():
         lista_paquetes.append('scipy')
 
     if len(lista_paquetes):
+
+        if not os.path.exists(directorio_móds):
+            dir_creado = True
+            os.makedirs(directorio_móds)
+        else:
+            dir_creado = False
+
         # Actualizar Pip
         _actualizar_pip()
 
@@ -127,7 +113,8 @@ def instalar_requísitos():
         for paq in lista_paquetes:
             _instalar_whl(paq)
 
-        shutil.rmtree('Módulos')
+        if dir_creado:
+            shutil.rmtree('Módulos')
 
     # Verificar que todo esté bien:
     try:
@@ -136,7 +123,7 @@ def instalar_requísitos():
         import matplotlib as _
     except ImportError:
         _ = None
-        raise ImportError('Error: No se instalaron todos los módulos necesarios.')
+        avisar('Error: No se instalaron todos los módulos necesarios.')
 
     try:
         import scipy.stats as _
@@ -145,22 +132,45 @@ def instalar_requísitos():
 
     except ImportError:
         _ = None
-        raise ImportError('¡Error! Por experencia personal, probablemente es porque no instalaste la versión del'
-                          '"Microsoft C++ 2015 redistributable" {}.'.format('x86' if bits == '32' else 'x64')
-                          )
+        avisar('¡Error! Por experencia personal, probablemente es porque no instalaste la versión del'
+               '"Microsoft C++ 2015 redistributable" {}.'.format('x86' if bits == '32' else 'x64')
+               )
 
-instalar_requísitos()
+
+if so == 'Windows':
+    sistema = 'win' + bits
+
+    for paquete, dic_paq in info_paquetes.items():
+        v = dic_paq['versión']
+        dic_paq['formato_archivo'] = dic_paq['formato_archivo'].format(versión=v, v_py=versión_python, sis=sistema)
+
+    instalar_requísitos()
+
+"""
+Ahora, las cosas normales de instalación.
+"""
+
+with open('tinamit/versión.txt') as archivo_versión:
+    versión = archivo_versión.read().strip()
 
 setup(
     name='tinamit',
     version=versión,
     packages=find_packages(),
-    url='https://github.com/julienmalard/Tinamit',
+    url='https://tinamit.readthedocs.io',
+    download_url='https://github.com/julienmalard/Tinamit',
     license='GNU GPL 3',
     author='Julien Jean Malard',
     author_email='julien.malard@mail.mcgill.ca',
     description='Conexión de modelos socioeconómicos (dinámicas de los sistemas) con modelos biofísicos.',
+    long_description='Tinamït permite conectar modelos biofísicos con modelos socioeconómicos de dinámicas de los'
+                     'sistemas (MDS). Es muy útil para proyectos de modelización participativa, especialmente'
+                     'en proyectos de manejo del ambiente. El interaz gráfico traducible facilita la adopción por'
+                     'comunidades en cualquier parte del mundo.',
     requires=['numpy', 'matplotlib', 'scipy'],
+    classifiers=[
+        'Programming Language :: Python :: 3.5',
+    ],
     package_data={
         # Incluir estos documentos de los paquetes:
         '': ['*.txt', '*.vpm', '*.json', '*.png', '*.jpg', '*.gif'],
